refactor: remaining fixtures/tests requiring sqlalchemy moved into sqlalchemy directory.
