+++ tests/unit/conftest.py
@@ -17,74 +17,6 @@ if TYPE_CHECKING:
     from starlite import Starlite
     from starlite.types import HTTPResponseBodyEvent, HTTPResponseStartEvent, HTTPScope
 
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.authors import Author
-    from tests.utils.domain.authors import Service as AuthorService
-    from tests.utils.domain.books import Book
-
-
-@pytest.fixture(name="author_repository_type")
-def fx_author_repository_type(
-    authors: list[Author], monkeypatch: pytest.MonkeyPatch
-) -> type[GenericMockRepository[Author]]:
-    """Mock Author repository, pre-seeded with collection data."""
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.authors import Author
-    from tests.utils.domain.authors import Service as AuthorService
-
-    repo = GenericMockRepository[Author]
-    repo.seed_collection(authors)
-    monkeypatch.setattr(AuthorService, "repository_type", repo)
-    return repo
-
-
-@pytest.fixture(name="author_repository")
-def fx_author_repository(
-    author_repository_type: type[GenericMockRepository[Author]],
-) -> GenericMockRepository[Author]:
-    """Mock Author repository instance."""
-    return author_repository_type()
-
-
-@pytest.fixture(name="author_service_type")
-def fx_author_service_type(_sqlalchemy_installed: None) -> type[AuthorService]:
-    from tests.utils.domain.authors import Service
-
-    return Service
-
-
-@pytest.fixture(name="book_repository_type")
-def fx_book_repository_type(
-    books: list[Book], monkeypatch: pytest.MonkeyPatch, _sqlalchemy_installed: None
-) -> type[GenericMockRepository[Book]]:
-    """Mock Book repository, pre-seeded with collection data."""
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.books import Book
-    from tests.utils.domain.books import Service as BookService
-
-    class BookRepository(GenericMockRepository[Book]):
-        """Mock book repo."""
-
-        model_type = Book
-
-    BookRepository.seed_collection(books)
-    monkeypatch.setattr(BookService, "repository_type", BookRepository)
-    return BookRepository
-
-
-@pytest.fixture(name="book_repository")
-def fx_book_repository(
-    book_repository_type: type[GenericMockRepository[Book]],
-) -> GenericMockRepository[Book]:
-    """Mock Book repo instance."""
-    return book_repository_type()
-
 
 @pytest.fixture()
 def http_response_start() -> HTTPResponseStartEvent:
